pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "release-notes-formatter"
version = "0.1.0"
description = "A tool for formatting release notes with Azure DevOps integration"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.2",
    "beautifulsoup4>=4.13.4",
    "azure-devops>=7.1.0b4",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "responses>=0.23.0",
    "pytest-cov>=6.1.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.setuptools.packages.find]
exclude = ["raw_release_notes"] 